feat: Use Babel Standalone for module transpilation
Switches the application to use Babel Standalone for client-side transpilation of ES modules. This change enables the use of newer JavaScript features that might not be natively supported by all target browsers, facilitating development and deployment. The script type in `index.html` has been updated to `text/babel` with `data-type="module"` to trigger this transpilation.